+++ projects/iplab/ngx-file-upload/src/lib/components/drop-zone/file-upload-drop-zone.component.ts
@@ -7,7 +7,8 @@ import {
     selector: `file-upload-drop-zone`,
     templateUrl: `./file-upload-drop-zone.component.html`,
     styleUrls: [`./file-upload-drop-zone.component.scss`],
-    changeDetection: ChangeDetectionStrategy.OnPush
+    changeDetection: ChangeDetectionStrategy.OnPush,
+    standalone: true
 })
 export class FileUploadDropZoneComponent {
 }