+++ projects/iplab/ngx-file-upload/src/lib/directives/discard.directive.ts
@@ -6,7 +6,8 @@ import { SimpleFileUploadComponent } from '../components/simple-file-upload/simp
 
 @Directive({
     selector: 'file-upload[discard]',
-    host: {'[attr.discard]': 'discard ? discard : null'}
+    host: {'[attr.discard]': 'discard ? discard : null'},
+    standalone: true
 })
 export class FilesDiscardDirective implements AfterViewInit, OnChanges {